requirements.txt
# Core LangChain
langchain==0.1.0
langchain-core==0.1.0
langchain-community>=0.0.30

# LLM Providers
langchain-openai>=0.1.0
langchain-google-genai>=0.0.12
langchain-xai>=0.0.1

# Vector Store & Embeddings
langchain-chroma>=0.1.0
chromadb>=0.4.25
sentence-transformers>=2.2.2
langchain-huggingface>=0.0.1

# Workflow
langgraph>=0.0.1

# Data Processing
langchain-text-splitters>=0.0.1
numpy>=1.24.3
scikit-learn>=1.3.2
scipy>=1.11.4

# Optional: For other embeddings and utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
